% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleGBM.R
\name{EnsembleGBM-get_lineage}
\alias{EnsembleGBM-get_lineage}
\title{Terminal Node Lineage}
\arguments{
\item{mod}{GBM model object}

\item{tree}{Tree number from the gbm model}

\item{child}{Local child position within the tree}
}
\value{
Returns a list of logical conditions than can be combined to create a
single rule.
}
\description{
Recursively build a rule starting from a terminal node and working backwards
to the root node. Logical statements are accumulated along the way.
}
