% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleGBM.R
\name{EnsembleGBM-make_node_map}
\alias{EnsembleGBM-make_node_map}
\title{Make Node Map}
\arguments{
\item{mod}{model object from gbm package
Creates look up lists that map unique terminal node IDs to gbm trees and
local node ids}
}
\description{
Make Node Map
}
