% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.rulefitFit}
\alias{importance.rulefitFit}
\title{Variable Importance}
\usage{
\method{importance}{rulefitFit}(x, newx, s = c("lambda.1se", "lambda.min"),
  ...)
}
\arguments{
\item{x}{rulefitFit object as output by the \code{\link{train}} function}

\item{newx}{development dataset usede to train the rulefit model}

\item{s}{lambda penalty term used to retrieve coefficients}
}
\description{
Variable Importance
}
