% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleFit_class.R
\docType{class}
\name{RuleFit-class}
\alias{RuleFit-class}
\alias{RuleFit}
\title{RuleFit Class}
\description{
RuleFit class that wraps a tree ensemble object and provides
methods for creating rules and training models with them.
}
\section{Fields}{

\describe{
\item{\code{ensemble}}{Wrapper class that provides access to the base ensemble model}

\item{\code{fit}}{glmnet model combining rules from ensemble}

\item{\code{rules}}{list of rules generated from the ensemble}

\item{\code{rids}}{rule ids. The integer identifier of the node from the ensemble.}

\item{\code{coefs}}{fitted model coefficients for the rule ensemble.}

\item{\code{intercept}}{model intercep from fitted model.}
}}


