% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleFit_class.R
\name{RuleFit-generate_sas_code}
\alias{RuleFit-generate_sas_code}
\title{Generate Rule Fit Model SAS Code}
\arguments{
\item{file}{Destination to print the SAS code}

\item{pfx}{Character prefix to prepend to rule code. No underscore needed.}

\item{dedup}{TRUE/FALSE for whether to combine rules that are exactly the
same. See details for more information}
}
\description{
Generate Rule Fit Model SAS Code
}
\details{
Depending on the tree ensemble used, it is possible to train a model
with duplicate rules. This can happen using GBM with a low shrinkage as trees
are more likely to be correlated under such conditions. The dedup parameter
will combine rules that are exactly the same using a hash function and sum
their coefficients to create single, aggregated rule.
}
