% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleFit_class.R
\name{rulefit}
\alias{rulefit}
\alias{rulefit.gbm}
\title{Create a Rule Fit Model}
\usage{
rulefit(x)

\method{rulefit}{gbm}(x)
}
\arguments{
\item{x}{a tree ensemble model}

\item{x}{a GBM model}
}
\value{
A RuleFit model object

A RuleFit model object
}
\description{
Create a Rule Fit Model using an tree ensemble as its foundation.

Create a Rule Fit Model from a GBM model
}
