% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleFit_class.R
\name{RuleFit-predict}
\alias{RuleFit-predict}
\title{Predict Rule Fit Model}
\arguments{
\item{x}{Dataset on which to make RuleFit model predictions.}

\item{kfold}{TRUE/FALSE for whether to return the out-of-fold predictions
from the trained glmnet model. A close estimate of how the model will
generalize.}
}
\value{
A numeric vector of predicted values.
}
\description{
Predict Rule Fit Model
}
