% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuleFit_class.R
\name{RuleFit-train}
\alias{RuleFit-train}
\title{Train Rule Fit Model}
\arguments{
\item{x}{Optional dataset to train the RuleFit model}

\item{y}{Dependent variable for training the model. Defaults to target used
to build the gbm model.}

\item{nfolds}{Number of cross validation folds to use for selecting the
optimal glmnet lambda. Defaults to 5.}

\item{alpha}{Regularization mixture term. Defaults to 1 for LASSO regression.
Can pass 0 for Ridge or (0, 1) for Elastic Net.}
}
\value{
A \code{rule_fit_modeL} object
}
\description{
Train a Rule Fit model from an existing ensemble model
}
